+++ src/frontend/TopBar.h
@@ -1,6 +1,6 @@
 /*
  *  OpenBangla Keyboard
- *  Copyright (C) 2016 Muhammad Mominul Huque <mominul2082@gmail.com>
+ *  Copyright (C) 2016-2018 Muhammad Mominul Huque <mominul2082@gmail.com>
  *
  *  This program is free software: you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -34,6 +34,7 @@ class LayoutViewer;
 class AboutDialog;
 class AboutFile;
 class SettingsDialog;
+class AutoCorrectDialog;
 
 class TopBar : public QMainWindow
 {
@@ -51,6 +52,7 @@ protected:
 private slots:
     void layoutMenuLayouts_clicked();
     void layoutMenuInstall_clicked();
+    void settingsMenuAutoCorrect_clicked();
     void settingsMenuFixedLayoutAutoVForm_clicked();
     void settingsMenuFixedLayoutAutoChandra_clicked();
     void settingsMenuFixedLayoutOldReph_clicked();
@@ -84,6 +86,7 @@ private:
     AboutDialog *aboutDialog;
     LayoutViewer *layoutViewer;
     SettingsDialog *settingsDialog;
+    AutoCorrectDialog *autoCorrectDialog;
 
     /* Layout Popup Menu */
     QMenu *layoutMenu;
@@ -93,6 +96,7 @@ private:
     QAction *layoutMenuInstall;
     /* Settings Popup Menu */
     QMenu *settingsMenu;
+    QAction *settingsMenuAutoCorrect;
     QAction *settingsMenuShowDialog;
     QMenu *settingsMenuFixedLayout;
     QAction *settingsMenuFixedLayoutAutoVForm;