+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 ## 1.3.0 - In development
+* OpenBangla Keyboard now features a Phonetic AutoCorrect Preference Dialog [#20](https://github.com/OpenBangla/OpenBangla-Keyboard/pull/20). User now can view/edit/add AutoCorrect entries.
 * Let users turn off candidate box [#11](https://github.com/OpenBangla/OpenBangla-Keyboard/issues/11)
 * Build system improvement. There is no need to pass `CMAKE_INSTALL_PREFIX` when building.
 * Now supports RPM packaging.